Extract postbuild scripts in separete file.
